Fix t\60-ignore-file.t test
The primary test case was never failing, regardless of the contents of
the cache. After fixing that, the test failed on Windows due to mixing
of '/' and '\' separators.

To fix the underlying Windows issue, add a new helper method
$m.path-for-fragment('some/fragment.pod6')
	=> '\path\to\docs\some\fragment.pod6'

Use that when registering .ignore-cache entries. It could be a private
method, but it's helpful for testing, too, so I've left it public. It
could be a subroutine, but it's always going to be relative to
$!doc-source so it makes sense as an instance method.