Changing how cache works... so that it *really* works
Main thing here is that we use the "ID" as a kind of "checksum" of what's in the file. The ID is computed as a sha1 (internally, we don't see that) of the name of the file "|" content of the file. Since the ID changes when any of them changes, that triggers the re-loading of the Cache

This closes #10, but I still have to do more extensive testing to see if nothing else breaks. Also moved 20-refresh to a subtest in 10-cache, since tests need to be independent. Still this one needs to be addressed.